+++ src/libtcod/txtfield.h
@@ -49,7 +49,11 @@ TCODLIB_API void TCOD_text_set_pos(TCOD_text_t txt, int x, int y);
 TCODLIB_API void TCOD_text_set_properties(
     TCOD_text_t txt, int cursor_char, int blink_interval, const char* prompt, int tab_size);
 TCODLIB_API void TCOD_text_set_colors(TCOD_text_t txt, TCOD_color_t fore, TCOD_color_t back, float back_transparency);
+
+TCODLIB_BEGIN_IGNORE_DEPRECATIONS
 TCODLIB_API bool TCOD_text_update(TCOD_text_t txt, TCOD_key_t key);
+TCODLIB_END_IGNORE_DEPRECATIONS
+
 TCODLIB_API void TCOD_text_render(TCOD_text_t txt, TCOD_console_t con);
 TCODLIB_API const char* TCOD_text_get(TCOD_text_t txt);
 TCODLIB_API void TCOD_text_reset(TCOD_text_t txt);